# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "StarWars API",
        "contact": {
            "name": "Sam Olabamiji",
            "email": "samolabams@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
        "/movies/{movieId}/characters": {
            "get": {
                "tags": [
                    "Character"
                ],
                "summary": "Get Movie Characters",
                "operationId": "getMovieCharacters",
                "parameters": [
                    {
                        "name": "movieId",
                        "in": "path",
                        "description": "The movie id parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "The filter query: to filter by gender (male / female)",
                        "schema": {
                            "type": "string"
                        },
                        "example": "female"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "The sort query: to sort by name, gender or height (add -sort to sort in descending order. Example: -name)",
                        "schema": {
                            "type": "string"
                        },
                        "example": "-height"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Characters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Character"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: If the service is not available to process request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found: If the movie is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error404"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{movieId}/characters/{characterId}": {
            "get": {
                "tags": [
                    "Character"
                ],
                "summary": "Get Movie Comments By Comment Id",
                "operationId": "getMovieCharacterByCharacterId",
                "parameters": [
                    {
                        "name": "movieId",
                        "in": "path",
                        "description": "The movie id parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "characterId",
                        "in": "path",
                        "description": "The character id parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Character"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: If the service is not available to process request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found: If the movie and/or comment is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error404"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Character": {
                "required": [
                    "id",
                    "name",
                    "gender",
                    "height"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Kevin hart"
                    },
                    "gender": {
                        "type": "string",
                        "example": "male"
                    },
                    "height": {
                        "type": "integer",
                        "format": "int64",
                        "example": "176"
                    }
                },
                "type": "object"
            },
...
            "Error500": {
                "required": [
                    "http_code",
                    "message"
                ],
                "properties": {
                    "http_code": {
                        "type": "integer",
                        "format": "int32",
                        "example": "500"
                    },
                    "message": {
                        "type": "string",
                        "example": "Internal Server Error"
                    }
                },
                "type": "object"
            },
            "Error404": {
                "required": [
                    "http_code",
                    "message"
                ],
                "properties": {
                    "http_code": {
                        "type": "integer",
                        "format": "int32",
                        "example": "404"
                    },
                    "message": {
                        "type": "string",
                        "example": "The requested resource is not available"
                    }
                },
                "type": "object"
            },
...
        }
    }
}